AP_Motors: rename FrameClass to MotorFrame
