+++ libraries/AP_Motors/AP_Motors6DOF.h
@@ -7,7 +7,7 @@
 #include <AP_Math/AP_Math.h>        // ArduPilot Mega Vector/Matrix math Library
 #include <RC_Channel/RC_Channel.h>     // RC Channel Library
 #include "AP_MotorsMatrix.h"
-#include "FrameClass_Sub.h"
+#include "MotorFrame_Sub.h"
 
 /// @class      AP_MotorsMatrix
 class AP_Motors6DOF : public AP_MotorsMatrix {
