+++ libraries/AP_Motors/MotorFrame.h
